@@ -11,9 +11,9 @@ import java.util.concurrent.Executor;
 @EnableAsync
 public class AsyncConfig {
 
-    @Bean(name ="taskExecutor")
-    public Executor taskExecutor(){
-        ThreadPoolTaskExecutor executor=new ThreadPoolTaskExecutor();
+    @Bean(name = "taskExecutor")
+    public Executor taskExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
         executor.setCorePoolSize(2);
         executor.setMaxPoolSize(2);
         executor.setQueueCapacity(100);
