@@ -3,5 +3,5 @@ package com.javatechie.executor.api.repository;
 import com.javatechie.executor.api.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
